face down creature spells include disguise, not just morph
